# docs/openapi/mdm-openapi1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MDM Device Posture API",
    "version": "1.1.0",
    "description": "Canonical API contract for MDM  device posture system. This JSON spec covers agent ingestion, normalization, evaluation, scoring, decisioning, remediation, policy management, catalog, lookups, and admin jobs."
  },
  "servers": [
    {
      "url": "https://mdm.24onlinebilling.com",
      "description": "Production"
    },
    {
      "url": "https://staging-mdm.24onlinebilling.com",
      "description": "Staging"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/policies/system-rules": {
      "get": {
        "tags": [
          "Policies"
        ],
        "summary": "List system rules",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Size"
          },
          {
            "in": "query",
            "name": "os_type",
            "schema": {
              "$ref": "#/components/schemas/OsType"
            }
          },
          {
            "in": "query",
            "name": "os_name",
            "schema": {
              "$ref": "#/components/schemas/LinuxOsName"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "$ref": "#/components/schemas/RecordStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "System rule list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Policies"
        ],
        "summary": "Create system rule",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemInformationRuleUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemInformationRule"
                }
              }
            }
          }
        }
      }
    },
    "/v1/policies/system-rules/{system_rule_id}": {
      "get": {
        "tags": [
          "Policies"
        ],
        "summary": "Get system rule",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/SystemRuleId"
          }
        ],
        "responses": {
          "200": {
            "description": "System rule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemInformationRule"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Policies"
        ],
        "summary": "Replace system rule",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/SystemRuleId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemInformationRuleUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemInformationRule"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Policies"
        ],
        "summary": "Patch system rule",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/SystemRuleId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemInformationRuleUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemInformationRule"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Policies"
        ],
        "summary": "Soft delete system rule",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/SystemRuleId"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        }
      }
    },
...
    "/v1/policies/system-rules/{system_rule_id}/conditions": {
      "get": {
        "tags": [
          "Policies"
        ],
        "summary": "List system rule conditions",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/SystemRuleId"
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Size"
          }
        ],
        "responses": {
          "200": {
            "description": "Condition list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Policies"
        ],
        "summary": "Create system rule condition",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/SystemRuleId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemInformationRuleConditionUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemInformationRuleCondition"
                }
              }
            }
          }
        }
      }
    },
    "/v1/policies/system-rules/{system_rule_id}/conditions/{condition_id}": {
      "get": {
        "tags": [
          "Policies"
        ],
        "summary": "Get condition",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/SystemRuleId"
          },
          {
            "$ref": "#/components/parameters/ConditionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Condition",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemInformationRuleCondition"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Policies"
        ],
        "summary": "Replace condition",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/SystemRuleId"
          },
          {
            "$ref": "#/components/parameters/ConditionId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemInformationRuleConditionUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemInformationRuleCondition"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Policies"
        ],
        "summary": "Patch condition",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/SystemRuleId"
          },
          {
            "$ref": "#/components/parameters/ConditionId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemInformationRuleConditionUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemInformationRuleCondition"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Policies"
        ],
        "summary": "Soft delete condition",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/SystemRuleId"
          },
          {
            "$ref": "#/components/parameters/ConditionId"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        }
      }
    },
    "/v1/policies/reject-apps": {
      "get": {
        "tags": [
          "Policies"
        ],
        "summary": "List reject app policies",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Size"
          }
        ],
        "responses": {
          "200": {
            "description": "Reject app policy list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Policies"
        ],
        "summary": "Create reject app policy",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectApplicationUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectApplicationPolicy"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/policies/trust-score-policies": {
      "get": {
        "tags": [
          "Policies"
        ],
        "summary": "List trust score policies",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Size"
          }
        ],
        "responses": {
          "200": {
            "description": "Trust score policy list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Policies"
        ],
        "summary": "Create trust score policy",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrustScorePolicyUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustScorePolicy"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "XTenantId": {
        "in": "header",
        "name": "X-Tenant-Id",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "Page": {
        "in": "query",
        "name": "page",
        "schema": {
          "type": "integer",
          "minimum": 0,
          "default": 0
        }
      },
      "Size": {
        "in": "query",
        "name": "size",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 20
        }
      },
...
      "SystemRuleId": {
        "in": "path",
        "name": "system_rule_id",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
...
    },
...
    "schemas": {
...
      "PagedEnvelope": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "page": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "size": {
                "type": "integer"
              },
              "total_elements": {
                "type": "integer"
              },
              "total_pages": {
                "type": "integer"
              }
            }
          }
        }
      },
      "OsType": {
        "type": "string",
        "enum": [
          "ANDROID",
          "IOS",
          "WINDOWS",
          "MACOS",
          "LINUX"
        ]
      },
      "LinuxOsName": {
        "type": "string",
        "enum": [
          "DEBIAN",
          "CENTOS",
          "FEDORA",
          "MINT",
          "REDHAT",
          "UBUNTU",
          "ROCKY",
          "ALMALINUX",
          "SUSE",
          "OTHER"
        ]
      },
...
      "RecordStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "INACTIVE"
        ]
      },
...
    }
  }
}